process at cregistration

+++ code/CLayer.h
@@ -27,20 +27,20 @@ class CLayer
   std::vector<CPoint*> vec_medial_points;
   CPoint* center_point;
   CPoint* moment_one_point;
-  double max_x,max_y,min_x,min_y;
+  float max_x,max_y,min_x,min_y;
 
   void reset();
   std::map<int,int> map_contourID;//contour_Num,contourID
   std::map<int,CContour*> map_contour;//contourID,CContour*
-  double** medial_axis;
+  float** medial_axis;
   int medial_axis_count;
   float total_length;
 private:
-  void check_edge(double,double);
+  void check_edge(float,float);
   void get_center();
   void check_length(float);
   void calculate_one_moment();
   float sum_x,sum_y;
-  double last_x,last_y;
+  float last_x,last_y;
   void calculate_medial_axis();
 };
