+++ code/CMedialMap.h
@@ -8,10 +8,10 @@ class CMedialMap:public CMap
 {
   public:
   //map is the medial map of the layer
-  CMedialMap(CContour* temp_contour,double** map):CMap(temp_contour)
+  CMedialMap(CContour* temp_contour,float** map):CMap(temp_contour)
   {
     medial_axis=map;
   }
   virtual void fillupoutside();
-  double** medial_axis;
+  float** medial_axis;
 };
