+++ code/CRegistration.h
@@ -18,11 +18,11 @@ class ParamRecord
     lattice_y=0;
     errorE=0;
   }
-  double **lattice_x;
-  double **lattice_y;
-  double errorE;
+  float **lattice_x;
+  float **lattice_y;
+  float errorE;
   void operator=(const ParamRecord &);
-  void initial(double errorE,CContour* higher);
+  void initial(float errorE,CContour* higher);
 };
 
 class CRegistration
@@ -40,26 +40,26 @@ class CRegistration
   int narrow_band;
   int vicinity_points_num;
   //compute the intensity of the sample points in distance map
-  void compute_intensity(std::vector<CPoint*>&,double**&,std::vector<double>&);
+  void compute_intensity(std::vector<CPoint*>&,float**&,std::vector<float>&);
   //compute the intensity of single point in distance map
-  double compute_intensity_by_point(double**& dist,double tx, double ty);
+  float compute_intensity_by_point(float**& dist,float tx, float ty);
   //the grill in distance map
   void init_lattice(CContour*);
   
-  double* xvb;//the lattice's x value
-  double* yvb;//the lattice's y value
+  float* xvb;//the lattice's x value
+  float* yvb;//the lattice's y value
   //bspline_update
   void bspline_update(CContour*,int mode,std::vector<CPoint*>&,std::vector<CPoint*>&);// mode = 0,use the normal lattice, mode = 1 ,use the new lattice.p
   //compute energy
-  double energy_func_square_diff(CContour*);
+  float energy_func_square_diff(CContour*);
   //drive the eqution
-  void gradientdescent_smoother(CContour* lower,CContour*higher,double errorE);
+  void gradientdescent_smoother(CContour* lower,CContour*higher,float errorE);
   //calculate the lattice
-  void calculate_dlattice_by_point(double**&,double**&,CContour*,CContour*);
+  void calculate_dlattice_by_point(float**&,float**&,CContour*,CContour*);
   
-  bool update_lattice(double**&,double**&m,CContour*,CContour*,double&);
+  bool update_lattice(float**&,float**&m,CContour*,CContour*,float&);
   ParamRecord current_params;
-  double cubic_spline(double u,int o);
+  float cubic_spline(float u,int o);
   void get_correspondence(CCorrespond*,CContour*higher,CContour*lower);
   int get_closest_point(std::vector<CPoint*>&,CPoint);
   int get_closest_point(std::vector<CPoint*>&vec_points,std::vector<CPoint*>&medial_points,CPoint,bool&);
@@ -68,18 +68,18 @@ class CRegistration
   void fill_the_hole(CCorrespond* corres,int& count,int first,int size,int last_index,int index, CPoint* point1, CContour *higher,int gap,bool isforeward);
   int get_relate_index(int first, int size, int old_index);
   void get_gap(int first, int size, int index, int last_index,int &real_gap,bool& isforeward );
-  double** XB;// original control point coordinates;
-  double** YB;
-  double** dXB;
-  double** dYB;
+  float** XB;// original control point coordinates;
+  float** YB;
+  float** dXB;
+  float** dYB;
 
   float lamda;// step size
   float kappa;//weight factor controlling the smoothness term
 
   const int NumberRows,NumberCols;
-  inline double max(double* val,int size)
+  inline float max(float* val,int size)
   {
-    double max = val[0];
+    float max = val[0];
     for (int i = 0; i < size; ++i)
     {
       if (val[i]>max)
@@ -89,9 +89,9 @@ class CRegistration
     }
     return max;
   }
-  inline double min(double*val,int size)
+  inline float min(float*val,int size)
   {
-    double min =val[0];
+    float min =val[0];
     for (int i = 0; i < size; ++i)
     {
       if (val[i]<min)
@@ -101,27 +101,27 @@ class CRegistration
     }
     return min;
   }
-  inline double dot(double*v1,double*v2)
+  inline float dot(float*v1,float*v2)
   {
     return v1[0]*v2[0]+v1[1]*v2[1];
   }
-  inline double spline_deriv(double u, int o)
+  inline float spline_deriv(float u, int o)
   {
     
-	double b;
+	float b;
 	switch (o)
 	{
 	case 0:
-		b = (double) 1.0/ (double) 6.0*3.0*(1.0-u)*(1.0-u)*(-1);
+		b = (float) 1.0/ (float) 6.0*3.0*(1.0-u)*(1.0-u)*(-1);
 		break;
 	case 1:
-		b = (double) 1.0/(double) 6.0*(3.0*3.0*u*u-6.0*2.0*u);
+		b = (float) 1.0/(float) 6.0*(3.0*3.0*u*u-6.0*2.0*u);
 		break;
 	case 2:
-		b = (double) 1.0/(double) 6.0*(-3.0*3.0*u*u+3.0*2.0*u+3.0);
+		b = (float) 1.0/(float) 6.0*(-3.0*3.0*u*u+3.0*2.0*u+3.0);
 		break;
 	case 3:
-		b = (double)1.0/(double) 6.0*3.0*u*u;
+		b = (float)1.0/(float) 6.0*3.0*u*u;
 		break;
 	}
 	return b;
